Fix: 정답률 로그 Int에서 Double로 수정

+++ backend/log-service/src/main/java/com/helloword/logservice/domain/log/repository/LogRepositoryCustomImpl.java
@@ -111,10 +111,10 @@ public class LogRepositoryCustomImpl implements LogRepositoryCustom {
 	}
 
 	@Override
-	public Integer findKidAverageCorrectRateByGameType(Long kidId, GameType gameType) {
+	public Double findKidAverageCorrectRateByGameType(Long kidId, GameType gameType) {
 		QLog log = QLog.log;
 
-		return queryFactory.select(log.correctRate.avg().intValue())
+		return queryFactory.select(log.correctRate.avg())
 			.from(log)
 			.where(log.kidId.eq(kidId)
 				.and(log.gameType.eq(gameType)))
@@ -122,10 +122,10 @@ public class LogRepositoryCustomImpl implements LogRepositoryCustom {
 	}
 
 	@Override
-	public Integer findGlobalAverageCorrectRateByGameType(GameType gameType) {
+	public Double findGlobalAverageCorrectRateByGameType(GameType gameType) {
 		QLog log = QLog.log;
 
-		return queryFactory.select(log.correctRate.avg().intValue())
+		return queryFactory.select(log.correctRate.avg())
 			.from(log)
 			.where(log.gameType.eq(gameType))
 			.fetchOne();
